shared: split platform parts into separate files
Reduce the ifdefy inside a file by splitting the platform code into
separate files.
