+++ shared/net-util-linux.c
